refactor(test): move code into `test_regex_database.py`
